[DFCT0010103] Fix blob deletion bug and allow cron job configuration (#658)
* change cron job always enabled to be configurable

* add env var to disable workflow cleanup

* add tests

* chore: lint

* chore: more lint

* fix test setting conf

* Fix blob deletion bug

* add configurable cron delete delay

* also check for empty string

* restructure cancel service code flow for simplicity

* update codeflow diagram

* fix unneeded check

* Add edge case tests

* requested changes

* upgrade pnpm github action from v2 to v4

* Fix code flow issue caught by test

+++ app/services/workflow_executions/cleanup_service.rb
@@ -9,9 +9,12 @@ module WorkflowExecutions
     end
 
     def execute
-      return if @workflow_execution.cleaned?
+      return if @workflow_execution.nil? || @workflow_execution.cleaned?
 
-      ActiveStorage::Blob.service.delete_prefixed(@workflow_execution.blob_run_directory)
+      # This check is for safety as passing nil or an empty string into deleted_refixed will delete all blobs
+      if @workflow_execution.blob_run_directory.present?
+        ActiveStorage::Blob.service.delete_prefixed(@workflow_execution.blob_run_directory)
+      end
 
       @workflow_execution.cleaned = true
